Version 2.0b - Wrapping up -
[ chat.kv ]
Fixed a bug where the tk window would stay open on app crash

[ client.py ]
Changed the way files are handled, now they are loaded only upon validation that the server is ready, and stored into the database.
Removed the need for a "file in queue", replaced with packet metadata
Changed the packet structure a bit to now also include the filename, in encrypted form
Changed when file is saved to DB and displayed to the user
Added partial decryption on ingest, to get the filename. This is faster than decrypting the whole file every time
Greatly improved speed and memory usage when dealing with multiple files per conversation, as they are now only loaded when their corresponding button is pressed.
Minor formatting.

[ Client/DBHandler.py ]
Updated the database structure to also include filename ( if applicable )
Added method for getting file data from database
Added method to only retrieve filename
Added extra parameter to save_message, this is a stopgap measure, might remove it in the future.
Minor reformatting.

[ UIElements.py ]
Added a JIT-like handling of files, decryption is done only when file is requested for download.
Added some extra metadata to the builder classes

[ server.py ]
Reply packets contain more metadata now, to accommodate changes made to the file transfer mechanism
Fixed cached messages not working properly, because buffer data wasn't shared between the clients' threads
Minor reformatting.

[ Server/DBHandler.py ]
Updated the structure of the database, to now contain the encrypted form of filenames, for partial decryption in client
Added extra checks to the cache functions, to mitigate some exceptions
Minor reformatting.